requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
ca-certificates=2020.10.14=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=pyhd3eb1b0_3
chardet=3.0.4=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py38_0
dbus=1.13.18=hb2f20db_0
expat=2.2.10=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
gast=0.3.3=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.33.2=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
intel-openmp=2020.2=254
joblib=0.17.0=pypi_0
jpeg=9b=h024ee3a_2
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.0=py38h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pypi_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
npy-append-array=0.9.6=pypi_0
numpy=1.18.5=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1h=h7b6447c_0
opt-einsum=3.3.0=pypi_0
pcre=8.44=he6710b0_0
pillow=8.0.1=py38he98fc37_0
pip=20.0.2=py38_3
protobuf=3.13.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
pyts=0.7.0=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.3=pypi_0
setuptools=50.3.0=py38hb0f4dca_1
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow=2.3.1=pypi_0
tensorflow-estimator=2.3.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
tornado=6.0.4=py38h7b6447c_1
tqdm=4.50.2=py_0
urllib3=1.25.11=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
